+++ src/app/app.module.ts
@@ -39,7 +39,6 @@ import { FooterComponent } from './main/footer/footer.component';
 import { GridComponent } from './main/grid/grid.component';
 import { HeaderComponent } from './main/header/header.component';
 import { HelpComponent } from './main/help/help.component';
-import { LoginComponent } from './main/login/login.component';
 import { MainComponent } from './main/main.component';
 import { FormattedBooleanPipe } from './main/pipes/formatting/formatted-boolean.pipe';
 import { KnoraDatePipe } from './main/pipes/formatting/knoradate.pipe';
@@ -50,16 +49,7 @@ import { TruncatePipe } from './main/pipes/string-transformation/truncate.pipe';
 import { SelectLanguageComponent } from './main/select-language/select-language.component';
 import { DatadogRumService } from './main/services/datadog-rum.service';
 import { MaterialModule } from './material-module';
-import { AddressTemplateComponent } from './project/board/address-template/address-template.component';
-import { AttributionTabViewComponent } from './project/board/attribution-tab-view/attribution-tab-view.component';
 import { BoardComponent } from './project/board/board.component';
-import { ContactsTabViewComponent } from './project/board/contacts-tab-view/contacts-tab-view.component';
-import { DatasetTabViewComponent } from './project/board/dataset-tab-view/dataset-tab-view.component';
-import { OrganisationTemplateComponent } from './project/board/organisation-template/organisation-template.component';
-import { PersonTemplateComponent } from './project/board/person-template/person-template.component';
-import { ProjectTabViewComponent } from './project/board/project-tab-view/project-tab-view.component';
-import { TermsTabViewComponent } from './project/board/terms-tab-view/terms-tab-view.component';
-import { UrlTemplateComponent } from './project/board/url-template/url-template.component';
 import { AddUserComponent } from './project/collaboration/add-user/add-user.component';
 import { CollaborationComponent } from './project/collaboration/collaboration.component';
 import { SelectGroupComponent } from './project/collaboration/select-group/select-group.component';
@@ -69,8 +59,6 @@ import { ListItemFormComponent } from './project/list/list-item-form/list-item-f
 import { ListItemComponent } from './project/list/list-item/list-item.component';
 import { ListComponent } from './project/list/list.component';
 import { OntologyFormComponent } from './project/ontology/ontology-form/ontology-form.component';
-import { OntologyVisualizerComponent } from './project/ontology/ontology-visualizer/ontology-visualizer.component';
-import { VisualizerComponent } from './project/ontology/ontology-visualizer/visualizer/visualizer.component';
 import { OntologyComponent } from './project/ontology/ontology.component';
 import { PropertyFormComponent } from './project/ontology/property-form/property-form.component';
 import { PropertyInfoComponent } from './project/ontology/property-info/property-info.component';
@@ -105,6 +93,7 @@ import { AddValueComponent } from './workspace/resource/operations/add-value/add
 import { DisplayEditComponent } from './workspace/resource/operations/display-edit/display-edit.component';
 import { PropertiesComponent } from './workspace/resource/properties/properties.component';
 import { AddRegionFormComponent } from './workspace/resource/representation/add-region-form/add-region-form.component';
+import { ArchiveComponent } from './workspace/resource/representation/archive/archive.component';
 import { AudioComponent } from './workspace/resource/representation/audio/audio.component';
 import { DocumentComponent } from './workspace/resource/representation/document/document.component';
 import { StillImageComponent } from './workspace/resource/representation/still-image/still-image.component';
@@ -177,13 +166,12 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         AccountComponent,
         AddGroupComponent,
         AddRegionFormComponent,
-        AddressTemplateComponent,
         AddUserComponent,
         AddValueComponent,
         AdminImageDirective,
         AdvancedSearchComponent,
         AppComponent,
-        AttributionTabViewComponent,
+        ArchiveComponent,
         AudioComponent,
         BoardComponent,
         BooleanValueComponent,
@@ -195,10 +183,8 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         ComparisonComponent,
         ConfirmationDialogComponent,
         ConfirmationMessageComponent,
-        ContactsTabViewComponent,
         CookiePolicyComponent,
         DashboardComponent,
-        DatasetTabViewComponent,
         DateEditComponent,
         DateInputComponent,
         DateInputTextComponent,
@@ -241,25 +227,20 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         ListItemFormComponent,
         ListValueComponent,
         ListViewComponent,
-        LoginComponent,
         LoginFormComponent,
         MainComponent,
         MembershipComponent,
         MessageComponent,
         OntologyComponent,
         OntologyFormComponent,
-        OntologyVisualizerComponent,
-        OrganisationTemplateComponent,
         PasswordFormComponent,
         PermissionComponent,
-        PersonTemplateComponent,
         ProfileComponent,
         ProgressIndicatorComponent,
         ProjectComponent,
         ProjectFormComponent,
         ProjectsComponent,
         ProjectsListComponent,
-        ProjectTabViewComponent,
         PropertiesComponent,
         PropertyFormComponent,
         PropertyInfoComponent,
@@ -302,7 +283,6 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         SublistValueComponent,
         SwitchPropertiesComponent,
         SystemComponent,
-        TermsTabViewComponent,
         TextValueAsHtmlComponent,
         TextValueAsStringComponent,
         TextValueAsXMLComponent,
@@ -312,13 +292,11 @@ export function httpLoaderFactory(httpClient: HttpClient) {
         TruncatePipe,
         UploadComponent,
         UriValueComponent,
-        UrlTemplateComponent,
         UserComponent,
         UserFormComponent,
         UserMenuComponent,
         UsersComponent,
         UsersListComponent,
-        VisualizerComponent,
         YetAnotherDateValueComponent,
     ],
     imports: [