+++ src/app/workspace/resource/properties/properties.component.ts
@@ -14,6 +14,7 @@ import {
     PermissionUtil,
     ProjectResponse,
     PropertyDefinition,
+    ReadColorValue,
     ReadLinkValue,
     ReadProject,
     ReadResource,
@@ -112,6 +113,8 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
      * can be used for preview when hovering on resource
      */
     @Output() referredResourceHovered: EventEmitter<ReadLinkValue> = new EventEmitter<ReadLinkValue>();
+
+    @Output() regionColorChanged: EventEmitter<ReadColorValue> = new EventEmitter<ReadColorValue>();
 
     lastModificationDate: string;
 
@@ -340,6 +343,7 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
                         this._dspApiConnection.v2.res.updateResourceMetadata(payload).subscribe(
                             (response: UpdateResourceMetadataResponse) => {
                                 this.resource.res.label = payload.label;
+                                this.lastModificationDate = response.lastModificationDate;
                             },
                             (error: ApiResponseError) => {
                                 this._errorHandler.showMessage(error);
@@ -355,8 +359,7 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
     /**
     * display message to confirm the copy of the citation link (ARK URL)
     */
-    openSnackBar() {
-        const message = 'ARK URL copied to clipboard!';
+    openSnackBar(message: string) {
         this._notification.openSnackBar(message);
     }
 
@@ -444,6 +447,9 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
                 });
             if (updatedValue instanceof ReadTextValueAsXml) {
                 this._updateStandoffLinkValue();
+            }
+            if (updatedValue instanceof ReadColorValue) {
+                this.regionColorChanged.emit();
             }
         } else {
             console.error('No properties exist for this resource');